mpi refactor step 2: encapsulate threads
